command line parser re-org in prep for logic update
The order of the switches is import.  They can be re-used.
For example, -R before the command verb refers to the installation root and
-R after the verb can be used again.

Moreover, aliases can't be supported with the way the parser is currently.
For now, organize the switches data structure to be cleaner and fix the
cascading errors.