two new queries
query 1) get list of files installed by a package (by ID)

query 2) super "which" query.
It builds a set of packages (limited to NSV, version data) but contains all the files that match the query path.
Normally this is one package with one file, but if globbing is used, it could be
multiple packages each with multiple files.