add man page for genrepo
This is the analog to "pkg repo".
The -o, -m, and -l options were dropped.
blake3 is used instead of sha256
the number of generated files was dropped to 2
There's a new concept of "reference" repository and mirrors (the catalog.sum can only come from the reference, but everything else can come from a mirror)