Add new option to genrepo
Implement the -f, --fingerprint option.
When used with --pubkey or --external, it creates a fingerprint configuration
file to be installed on the client at the given location.

moreover, if the public key is packaged with the catalog archive, the
fingerprint will be printed to the screen (if quiet option not set)