Add true sign server
I mimicked the process to set up a signed server as outlined by pkg man pages.
Then I realized that the process was flawed because it was being used to keep
the key secret, but any SSH user that can run ravensign can also just use SSH
to reveal the key.

So it needs a webserver running with permissions to be able to do it for us
to keep the key secure.

example of sign command:
ssh <host> curl -s 127.0.0.1:8080 --data digest=<64-character-hex-string>
This requires a small change in how genrepo -x inserts the digest in the command.