drop global switch--rootdir
This option has the potential to break a lot of crap.
The man issue is the database.  Does the database capture the rootdir or not?
What happens if some packages are installed with --rootdir and others aren't?
We don't have these issues with --chroot option and really there's no benefit to
using --rootdir over --chroot, so just limit our support to --chroot.