first attempt at detecting variant clash
for example, searching for "nginx~" will result in both variants of nginx being selected.
with the implicit assumption that variants will install files in the same place,
selecting 2 variants of the same port should almost be considered an error.

This could should list the clashes and halt the attempt with an error.
The error message can not be silenced with --quiet.

Requires testing.